--- android/app/src/main/res/drawable/splashscreen.xml
@@ -1,3 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-  <item android:drawable="@color/splashscreen_background"/>
-</layer-list>
