--- android/app/src/main/res/values/colors.xml
@@ -1,5 +1,0 @@
-<resources>
-  <color name="splashscreen_background">#ffffff</color>
-  <color name="colorPrimary">#023c69</color>
-  <color name="colorPrimaryDark">#ffffff</color>
-</resources>
